[Add]星表示、文字サイズ調整

+++ db/migrate/20221106061806_create_reviews.rb
@@ -1,12 +1,12 @@
 class CreateReviews < ActiveRecord::Migration[6.1]
   def change
     create_table :reviews do |t|
-      
+
       t.integer :game_id, null: false
       t.integer :user_id, null: false
       t.string :review_title, null: false
       t.text :review_comment, null: false
-      t.float :rate, null: false
+      t.integer :rate, null: false
 
       t.timestamps
     end
